requirements.txt
Django==4.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-utils-six==2.0
